==> notebooks/1- Proyecto_limpieza_movies.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
sns.set()

data_movies = pd.read_csv('.\Datasets\movies_dataset.csv')


Limpieza de datos del dataset Movies

df_movies = pd.DataFrame(data_movies)

# Como mi sistema de recomendacion solo sera de pelis en español, primero elimino todas las filas cuyo
# lenguaje original no sea español

df_movies = df_movies[df_movies['original_language'] == "es"] 
df_movies.info()

# Rellenar con número 0 los valores nulos de los campos revenue y budget

# Chequeo primero si hay nulos y compruebo que no, por lo que no relleno nada
nulos_en_revenue = df_movies['revenue'].isnull().sum()
nulos_en_budget = df_movies['budget'].isnull().sum()

print(f"Cantidad de valores nulos en la columna 'revenue': {nulos_en_revenue}")
print(f"Cantidad de valores nulos en la columna 'budget': {nulos_en_budget}")

# Elimino Los valores nulos del campo release date 

# Chequeo primero si hay nulos en el campo indicado
nulos_en_release_date = df_movies['release_date'].isnull().sum()
print(f"Cantidad de valores nulos en la columna 'release_date': {nulos_en_release_date}")

# Como tengo dos nulos, procedo a eliminarlos
# Eliminar filas con valores nulos en la columna 'A'
df_movies = df_movies.dropna(subset=['release_date'])

# De haber fechas, deberán tener el formato AAAA-mm-dd
# Además deberán crear la columna release_year donde extraerán el año de la fecha de estreno y la columna 
# release_month


# Observo el dataset y compruebo que la única columna con fecha es "release_date" pero que ya está en ese formato
# df_movies.head()

#Para crear las columnas "release_year" y "release_month" debos convertirlas la columna original "release_date" a tipo datetime
# Como no me la convertía, busqué los campos que me lo impedían
df_movies['release_date'] = pd.to_datetime(df_movies['release_date'], errors='coerce')

# Verifico el tipo de datos de la columna 'release_date' y veo que ya es datetime
print(df_movies['release_date'].dtype)

# Extraigo el año y creo la nueva columna 'release_year' y hago lo mismo con el mes
df_movies['release_year'] = df_movies['release_date'].dt.year
df_movies['release_month'] = df_movies['release_date'].dt.month

df_movies.head()

# Crear la columna con el retorno de inversión, llamada return, con los campos revenue y budget, dividiendo estas dos últimas revenue / budget
# Cuando no hay datos disponibles para calcularlo, deberá tomar el valor 0.

# Como me tiraba error, chequeo que en las columnas revenue y budget no haya valores no numericos
revenue_no_numericos = df_movies[~df_movies['revenue'].apply(lambda x: isinstance(x, (int, float)))]
budget_no_numericos = df_movies[~df_movies['budget'].apply(lambda x: isinstance(x, (int, float)))]

print(f"Cantidad de valores nulos en la columna 'release_date': {revenue_no_numericos}")
print(f"Cantidad de valores nulos en la columna 'release_date': {budget_no_numericos}")


# Investigo qué tipo de dato tiene budget, porque todos son no numericos
df_movies["budget"].dtype


# Convierto la columna 'budget' a tipo numérico, forzando errores a NaN
df_movies['budget'] = pd.to_numeric(df_movies['budget'], errors='coerce')


# Verifico si hay valores NaN en la columna 'budget'
hay_nan = df_movies['budget'].isna().any()

print(f"¿Hay valores NaN en la columna 'budget'? {hay_nan}")

# Compruebo que cambió el tipo de dato
df_movies["budget"].dtype

# Ahora sí creo la nueva columna

df_movies["return"] = df_movies["revenue"] / df_movies["budget"]
df_movies.head()


# Cuento la cantidad de valores NaN en la columna 'return' para saber qué valor tendrá esta informacion
cantidad_nan_return = df_movies['return'].isna().sum()

print(f"Cantidad de valores NaN en la columna 'return': {cantidad_nan_return}")

# Antes de proceder a eliminar columnas chequeo si revenue, budget y return tienen muchos 0, porque si es así no me van a servir para mi modelo
# Contar la cantidad de valores 0 en la columna 'budget'
cantidad_zero_budget = (df_movies['budget'] == 0).sum()
print(f"Cantidad de valores 0 en la columna 'budget': {cantidad_zero_budget}")
cantidad_zero_revenue = (df_movies['revenue'] == 0).sum()
print(f"Cantidad de valores 0 en la columna 'revenue': {cantidad_zero_revenue}")
cantidad_zero_return = (df_movies['return'] == 0).sum()
print(f"Cantidad de valores 0 en la columna 'return': {cantidad_zero_return}")

# Inspecciono las columnas para tomar la decisión de eliminarlas
df_seleccion_columnas = df_movies[['budget', 'revenue', 'return']]
print(df_seleccion_columnas)

# Eliminar las columnas que no serán utilizadas: video,imdb_id,adult,original_title,poster_path y homepage.
# "Adult" la voy a dejar por las dudas porque puede ser importante en un sistema de recomendacion
# Y decido eliminar otras que tampoco me van a servir para el sistema de recomendacion

df_movies = df_movies.drop(columns=['belongs_to_collection', 'budget', 'homepage','imdb_id','original_language', 'original_title', 'poster_path','production_companies', 'release_date', 'revenue', 'runtime','spoken_languages', 'status','tagline', 'video'])

df_movies = df_movies.drop(columns=['return'])

df_movies.head()

# Exporto el DataFrame a un nuevo archivo CSV para, en adelante, trabajar con ese y no correr todo el codigo de nuevo
df_movies.to_csv('movies_dataset_limpio.csv', index=False)
