==> notebooks/2- Proyecto_desanidar_movies.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import json
sns.set()

data_movies_final = pd.read_csv('.\Datasets\movies_dataset_limpio.csv')



df_movies_final = pd.DataFrame(data_movies_final)

1- Voy a desanidar la columna genres para quedarme solo con los nombres de los géneros de cada película

# Imprimo para ver qué trae la columna genres
print(df_movies_final['genres'].head())

# Itero sobre cada fila para extraer solo el primer genero y agregar una columna con él

for index, row in df_movies_final.iterrows():
    genre = json.loads(row['genres'].replace("'", "\""))
    if genre:
        df_movies_final.at[index, 'genre'] = genre[0]['name']
    else:
        df_movies_final.at[index, 'genre'] = None



df_movies_final.head()

# Elimino la columna original "genres"
df_movies_final = df_movies_final.drop(columns=["genres"])


df_movies_final.head()

2- Ahora voy a desanidar la columna "production_countries" para quedarme solo con el país

# Iterar sobre cada fila del DataFrame

for index, row in df_movies_final.iterrows():
    country = json.loads(row['production_countries'].replace("'", "\""))
    if country:
        df_movies_final.at[index, 'country'] = country[0]['name']
    else:
        df_movies_final.at[index, 'country'] = None

df_movies_final.head()

# Elimino la columna original "production_countries"
df_movies_final = df_movies_final.drop(columns=["production_countries"])

# De la misma manera que antes, veo con cuántas columnas de país tiene sentido que me quede
df_movies_final.head()

3- El objetivo es trabajar el dataset Credits: reducirlo y que me queden solo las pelis de habla hispana, 
es decir las mismas que tengo en movies_final

# Levanto el csv Credits. Al levantarlo una segunda vez me tiró error en una línea, por eso
# decidí ignorar esa fila
data_credits = pd.read_csv('.\Datasets\credits.csv')



# Convierto a dataframe 
df_credits = pd.DataFrame(data_credits)


df_credits.info

# Filtro el dataset credits para quedarme solo con las filas cuyo id esté presente en el dataset data_movies_final
df_credits_filtered = df_credits[df_credits['id'].isin(df_movies_final['id'])]


# Compruebo que df_credits_filtered tiene una pelicula menos. 

df_credits_filtered.info()
df_movies_final.info()

# Voy a buscar cuál es para eliminarla
missing_movies = df_movies_final[~df_movies_final['id'].isin(df_credits_filtered['id'])]

# Mostrar las películas faltantes
print("Películas presentes en 'data_movies_final' pero no en 'credits_filtered':")
print(missing_movies)

# Elimino la fila con la película que no estaba en el dataset de Credits
df_movies_final = df_movies_final[df_movies_final['id'] != 401840]

4- Reducido el dataset de Credits, procedo a su limpieza

df_credits_filtered.head()

# Extraigo el nombre del primer actor que aparece en la lista de diccionarios de la columna cast
# Por algún motivo que no logré ubicar solo reconoce al actor principal en unas 100 de las 900 películas

# Función para extraer el nombre del primer actor
def get_first_actor(cast_str):
    try:
        cast_list = json.loads(cast_str.replace("'", "\""))
        if cast_list:
            return cast_list[0]['name']
        return None
    except json.JSONDecodeError:
        return None

# Aplicar la función a cada fila de la columna 'cast' y crear una nueva columna 'principal_cast'
df_credits_filtered['principal_cast'] = df_credits_filtered['cast'].apply(get_first_actor)



# Extraigo el nombre del director de la columna crew

# Función para extraer el nombre del director y el job
def get_director_info(crew_str):
    try:
        crew_list = json.loads(crew_str.replace("'", "\""))
        for member in crew_list:
            if member['job'] == 'Director':
                return pd.Series([member['job'], member['name']])
        return pd.Series([None, None])
    except json.JSONDecodeError:
        return pd.Series([None, None])

# Aplico la función a cada fila de la columna 'crew' y creo nuevas columnas 'job' y 'director'
df_credits_filtered[['job', 'director']] = df_credits_filtered['crew'].apply(get_director_info)

# Voy  a eliminar las columnas que no necesito
df_credits_filtered = df_credits_filtered.drop(columns=["cast", "crew", "job"])


# Concateno los dataframes por la columna 'id'
df_movies_final = pd.merge(df_movies_final, df_credits_filtered, on='id', how='inner')



df_movies_final.head()

# Exporto el dataframe concatenado en un nuevo archivo CSV
df_movies_final.to_csv('movies_dataset_sin_meses.csv', index=False)
