==> notebooks/3- Proyecto_creacion_funciones.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import json
sns.set()

### CREACIÓN DE FUNCIONES


##### FUNCIÓN CANTIDAD PELÍCULAS POR MES

Creo la función def cantidad_filmaciones_mes( Mes )

Se ingresa un mes en idioma Español. 

Debe devolver la cantidad de películas que fueron estrenadas en el mes consultado en la totalidad del dataset.

Ejemplo de retorno: X cantidad de películas fueron estrenadas en el mes de X


# Levanto el csv de producción y lo convierto en dataframe

data_movies_produccion = pd.read_csv('.\Datasets\movies_dataset_produccion.csv')
df_movies_produccion = pd.DataFrame(data_movies_produccion)

df_movies_produccion.head(10)

# Primero tengo que convertir la columna release_month (que está el mes en número) a nombre del mes

# Con este diccionario mapeeo los números de los meses a nombres en español
meses = {
    1: 'enero', 2: 'febrero', 3: 'marzo', 4: 'abril', 
    5: 'mayo', 6: 'junio', 7: 'julio', 8: 'agosto', 
    9: 'septiembre', 10: 'octubre', 11: 'noviembre', 12: 'diciembre'
}

# Creo una nueva columna con los nombres de los meses
df_movies_produccion['release_month_name'] = df_movies_produccion['release_month'].map(meses)

# Elimino la columna release_month pues ya no la usare
df_movies_produccion = df_movies_produccion.drop(columns=['release_month'])
    

def cantidad_filmaciones_mes(df):
    # Armo una lista de meses válidos en español para que luego me avierta si lo ingresado no es un mes
    meses_validos = [
        'enero', 'febrero', 'marzo', 'abril', 
        'mayo', 'junio', 'julio', 'agosto', 
        'septiembre', 'octubre', 'noviembre', 'diciembre'
    ]
    
    # Solicito al usuario que ingrese un mes
    mes = input("Por favor, ingrese un mes en español para saber cuántas películas fueron estrenadas en esa fecha: ").strip().lower()
    
    # Verifico si el mes ingresado es válido
    if mes not in meses_validos:
        return f"El mes '{mes}' no es válido. Por favor, ingresa un mes correcto."
    
    # Cuento las películas estrenadas en el mes especificado
    cantidad = df[df['release_month_name'].str.lower() == mes].shape[0]
    
    return f"{cantidad} películas fueron estrenadas en el mes de {mes}."

# Uso la función
cantidad_filmaciones_mes(df_movies_produccion)

##### FUNCIÓN INGRESAR TÍTULO Y RECIBIR AÑO DE ESTRENO Y POPULARIDAD

Creo la función def score_titulo( titulo_de_la_filmación ): 

Se ingresa el título de una filmación esperando como respuesta el título, el año de estreno y el score.

Ejemplo de retorno: La película X fue estrenada en el año X con un score/popularidad de X


df_movies_produccion.columns


def score_titulo(df):
    # Solicito al usuario que ingrese el título de la película
    titulo = input("Por favor, ingresa el título de una película de habla hispana: ").strip().lower()
    
    # Filtro el dataframe para encontrar la película por título
    pelicula = df[df['title'].str.lower() == titulo]
    
    # Advierto si no encuentra la película
    if pelicula.empty:
        return f"No se encontró ninguna película con el título '{titulo}'."
    
    # Obtengo los datos de la película
    titulo_pelicula = pelicula['title'].values[0]
    año_estreno = pelicula['release_year'].values[0]
    score = pelicula['popularity'].values[0]
    
    return f"La película '{titulo_pelicula}' fue estrenada en el año {año_estreno} con un score/popularidad de {score}."



# Uso la función
score_titulo(df_movies_produccion)

##### FUNCIÓN INGRESAR TÍTULO Y RECIBIR PROMEDIO VOTACIONES

Creo la función def votos_titulo( titulo_de_la_filmación )

Se ingresa el título de una filmación esperando como respuesta el título, la cantidad de votos y el valor promedio de las votaciones. 

La misma variable deberá de contar con al menos 2000 valoraciones, caso contrario, debemos contar con un mensaje avisando que no cumple esta condición y que por ende, no se devuelve ningún valor.

Ejemplo de retorno: La película X fue estrenada en el año X. La misma cuenta con un total de X valoraciones, con un promedio de X


# Exploro media, mediana y moda para encontrar mi filtro a la hora de considerar qué películas consiederar
# La consiga ponía ese filtro en 2000 valoraciones, pero mi dataset es de apenas 1000 peliculas
# Elijo las películas que tengan más de 10 valoraciones, es decir un poco más arriba de la mediana, para que entre casi la mitad de las peliculas del dataet

media = df_movies_produccion['vote_count'].mean()
mediana = df_movies_produccion['vote_count'].median()
moda = df_movies_produccion['vote_count'].mode()[0]

print(f"Media: {media}")
print(f"Mediana: {mediana}")
print(f"Moda: {moda}")

def votos_titulo(df):
    # Solicito al usuario que ingrese el título de la película
    titulo = input("Por favor, ingresa el título de una película de habla hispana: ").strip().lower()
    
    # Filtro el dataframe para encontrar la película por título
    pelicula = df[df['title'].str.lower() == titulo]
    
    # Advierto si no encuentra la película
    if pelicula.empty:
        return f"No se encontró ninguna película con el título '{titulo}'."
    
    # Obtengo los datos de la película
    titulo_pelicula = pelicula['title'].values[0]
    cantidad_votos = pelicula['vote_count'].values[0]
    
    # Verifico si la cantidad de votos es menor a 10
    if cantidad_votos < 10:
        return f"La película '{titulo_pelicula}' no cuenta con la siguiente cantidad de valoraciones para ser considerada."
    
    promedio_votos = pelicula['vote_average'].values[0]
    
    return f"La película '{titulo_pelicula}' recibió {cantidad_votos} votos y su promedio de valoración fue de {promedio_votos}."



# Uso la función
votos_titulo(df_movies_produccion)
