==> notebooks/4- Proyecto_conversion_release_month.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import json
sns.set()

movies_dataset_produccion = pd.read_csv(r'C:\Users\User\Documents\Henry\DataScience\Proyecto_Individual_1\Datasets\movies_dataset_produccion.csv', low_memory=False)


df_movies_produccion = pd.DataFrame(movies_dataset_produccion)

df_movies_produccion.head()


# Crear un diccionario de mapeo
month_mapping = {
    1: 'Enero', 2: 'Febrero', 3: 'Marzo', 4: 'Abril',
    5: 'Mayo', 6: 'Junio', 7: 'Julio', 8: 'Agosto',
    9: 'Septiembre', 10: 'Octubre', 11: 'Noviembre', 12: 'Diciembre'
}

# Reemplazar los valores en la misma columna 'release_month'
df_movies_produccion['release_month'] = df_movies_produccion['release_month'].map(month_mapping)




df_movies_produccion.head()

# Exporto el DataFrame a un nuevo archivo CSV para, en adelante, trabajar con ese y no correr todo el codigo de nuevo
df_movies_produccion.to_csv('movies_dataset_produccion.csv', index=False)
