==> notebooks/6- Proyecto_recomendacion.ipynb <==
import numpy as np
import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.preprocessing import MinMaxScaler



sns.set()

dataset_recomendacion = pd.read_csv(r'C:\Users\User\Documents\Henry\DataScience\Proyecto_Individual_1\Datasets\movies_dataset_produccion.csv', low_memory=False)

df_recomendacion = pd.DataFrame(dataset_recomendacion)

# Relleno los valores NaN con una cadena vacía
df_recomendacion = df_recomendacion.fillna('')


df_recomendacion.head()


# Función para recomendar películas basadas en la similitud del coseno
def recomendacion(title, df):
    # Inicializo el vectorizador TF-IDF
    tfidf = TfidfVectorizer(stop_words='english')
    
    # Ajusto y transformo las características combinadas
    tfidf_matrix = tfidf.fit_transform(df['overview'] + ' ' + df['genre'] + ' ' + df['country'] + df['vote_average'].astype(str))

    # Calculo la matriz de similitud del coseno
    cosine_sim = cosine_similarity(tfidf_matrix, tfidf_matrix)
    
    # Obtengo el índice de la película que coincide con el título
    idx = df.index[df['title'] == title].tolist()[0]
    
    # Obtengo los puntajes de similitud de todas las películas con esa película
    sim_scores = list(enumerate(cosine_sim[idx]))
    
    # Ordeno las películas según los puntajes de similitud
    sim_scores = sorted(sim_scores, key=lambda x: x[1], reverse=True)
    
    # Obtengo los puntajes de las 5 películas más similares
    sim_scores = sim_scores[1:6]
    
    # Obtengo los índices de las películas
    movie_indices = [i[0] for i in sim_scores]
    similarity_scores = [round(i[1], 5) for i in sim_scores]
    
    # Devuelvo las 5 películas más similares
    return df['title'].iloc[movie_indices], similarity_scores



# Obtengo todos los valores de la columna "title" como una lista para hacer pruebas con los títulos correctos 
title_list = df_recomendacion['title'].tolist()
title_list

# Ejemplo de uso
peliculas_recomendadas, score_similitud = recomendacion('El Amor Brujo', df_recomendacion)
print("Películas recomendadas y sus puntajes de similitud:")
for title, score in zip(peliculas_recomendadas, score_similitud):
    print(f"{title}: {score}")

recomendacion('Nueba Yol', df_recomendacion)
