==> notebooks/7- Proyecto_nan.ipynb <==
import numpy as np
import pandas as pd
from sklearn.linear_model import LinearRegression
from collections import Counter
from random import randint




movies_nan = pd.read_csv(r'C:\Users\User\Documents\Henry\DataScience\Primer_Proyecto\dataset\movies_dataset_produccion.csv', low_memory=False)

df = pd.DataFrame(movies_nan)

# Supongamos que df es tu DataFrame
print(df.isna().sum())  # Esto te mostrará cuántos valores NaN hay en cada columna

df['overview'].fillna('Sin info', inplace=True)
df['genre'].fillna('Sin info', inplace=True)
df['country'].fillna('Sin info', inplace=True)


# Supongamos que df es tu DataFrame
print(df.isna().sum())  # Esto te mostrará cuántos valores NaN hay en cada columna

df.to_csv('movies_dataset_produccion.csv', index=False)
